chanstate: move forwarding kv package
Move the current forwarding-package KV implementation into
chanstate/kv_forwarding_package.go and leave channeldb with the
temporary compatibility aliases used by existing callers.

This keeps the forwarding-package store as its own review slice and
leaves migration code untouched.
